Use ets:update_counter/4
ets:update_counter/4 lets us specify a default object to insert into
an ETS table if the counter we want to increase isn't yet present.  It
was introduced in Erlang/OTP 18.

This means we don't have to catch badarg errors.  Catching an error
with a plain 'catch' is inefficient starting in Erlang/OTP 15; see
http://erlang.org/pipermail/erlang-questions/2013-November/075928.html

Change statman_gauge:incr/2 to use a try/catch around
ets:update_counter/3: in this case we would need to always calculate
the current timestamp in order to pass it as a default object, and I'm
not sure whether that would be a net win.  Let's use a try-catch in
order to not build the stack trace, at least.